区域新增接口修改

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/domain/area/AreaVO.java
@@ -45,7 +45,7 @@ public class AreaVO extends AbstractObject {
      * 区域分类ID
      */
     @ApiModelProperty("区域分类ID")
-    private Integer areaTypeId;
+    private Long areaTypeId;
 
     /**
      * 区域名称
